refactor(provider/cf): move cf Job classes to correct modules to avoid siloed code
