+++ clouddriver-cloudfoundry/src/main/java/com/netflix/spinnaker/clouddriver/cloudfoundry/deploy/description/CloudFoundryRunJobOperationDescription.java
@@ -12,21 +12,31 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package com.netflix.spinnaker.clouddriver.cloudfoundry.job;
+package com.netflix.spinnaker.clouddriver.cloudfoundry.deploy.description;
 
-import com.netflix.spinnaker.clouddriver.cloudfoundry.deploy.description.AbstractCloudFoundryDescription;
+import com.netflix.frigga.Names;
 import com.netflix.spinnaker.clouddriver.cloudfoundry.model.CloudFoundryServerGroup;
+import com.netflix.spinnaker.clouddriver.security.resources.ApplicationNameable;
+import java.util.Collection;
+import java.util.Collections;
 import javax.annotation.Nullable;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
-public class CloudFoundryRunJobOperationDescription extends AbstractCloudFoundryDescription {
+public class CloudFoundryRunJobOperationDescription extends AbstractCloudFoundryDescription
+    implements ApplicationNameable {
 
   private CloudFoundryServerGroup serverGroup;
   @Nullable private String jobName;
   private String command;
+
+  @Override
+  public Collection<String> getApplications() {
+    return Collections.singletonList(Names.parseName(serverGroup.getName()).getApp());
+  }
 }
